created separate ts and scss files
