+++ src/index.html
@@ -31,7 +31,7 @@
       </nav>
       <div class="navigationlinks" id="navmenu">
         <ul class="navigationlinks_button">
-          <li><a href="#">Produkter</a></li>
+          <li><a href="#productsContainer">Produkter</a></li>
           <li><a href="#">Golfklubbor</a></li>
           <li><a href="#">Headcovers</a></li>
           <li><a href="#">Golfbollar</a></li>
